Adding methods to obtain more mesh information
Adding a serial method to obtain the information printed by
PrintCharacteristics.  This includes minimum and maximum edge lengths
as well as kappa which seems to be related to the aspect ratio of the
elements.

Adding a parallel method to compute the same mesh characteristics
mentioned above.

Adding a parallel method to compute the bounding box of a distributed
mesh.