Update the place where the headers installed - instead of installing {config,general,linalg,mesh,fem} in <PREFIX>/include, they are now installed in <PREFIX>/include/mfem. In addition, the master headers, mfem.hpp and mfem-performance.hpp, are also installed in <PREFIX>/include/mfem. However in order to keep the include lines    #include <mfem.hpp>    #include <mfem-performance.hpp> valid, the build system generates files mfem.hpp and mfem-performance.hpp in <PREFIX>/include that simply include the files with the same name from the mfem sub-directory.
This change should remain backward compatible with existing code, unless
the code included directly specific mfem headers like:
   #include "general/socketstream.hpp"
which will need to be updated to:
   #include "mfem/general/socketstream.hpp"

In addition, the files config.mk and test.mk are now installed in
<PREFIX>/share/mfem instead of <PREFIX>. Also, the location of test.mk
is now defined by config.mk, so that after including config.mk, the
file test.mk can be included (if needed) simply with:
   include $(MFEM_TEST_MK)

This layout is friendlier for installations with other packages in the
same installation prefix, e.g. /usr/local or $HOME/local.

TODO: mention these updates in CHANGELOG.