Fixes for *ArrayCoefficient classes
These classes take ownership of the Coefficients used for the i-th direction or (i,j) entry.
When a new one is set, the old one should be deleted
Added capability of having NULL coefficients, instead of having the user forced to create a Zero coefficient.
Changed signature of GetCoeff, which now returns a pointer (that can be NULL) instead of a reference