[tmop-dev] Various updates.
Add files generated by the new miniapps to .gitignore.

In class BilinearFormIntegrator, move the field IntRule and the method
SetIntRule to the base class, NonlinearFormIntegrator.

In class Vector, add method FNorm2 (Frobenius norm squared).

In class HyperelasticModel, update the implementations of some of the
static methods for evaluating the invariants and their derivatives.

In makefile, enhance the output of the 'print-%' target and add new
target, 'printall', that outputs all makefile variables.