Improves robustness of IsoparametricTransformation's TransformBack function.
The TransformBack function had an optimization to project points
that moved outside the element onto the element boundary. This caused
the code to incorrectly determine some internal points as outside on curved elements.

This commit removes this optimization and instead checks whether the found
IntegrationPoint is inside or outside the element once the algorithm converges.