Added overload with option to set refinement order in TransformBack
The refinement order is used to choose the initial IntegrationPoint
for the Newton-Raphson solve in IsoparametricTransformation's TransformBack
function. The default value is to use the elements order.

Also added a new public function to IsoparametricTransfomation to find
the closest IntegrationPoint within an element to a given point in space.
This function uses the elements RefinedIntules for the given order.