Vector::Norml2: use LAPACK dnrm2 algorithm
Use the LAPACK dnrm2 algorithm to compute the l2 norm instead of the
square root of the inner product of a vector with itself. This
modification prevents overflow by ensuring that any squared term has a
nonnegative argument of 1.0 or less. However, it also disables OpenMP
parallelism (because mfem::Vector::operator* is no longer being used);
this functionality would have to be reintroduced in a subsequent
commit.