In class BilinearFormIntegrator, implement the method AssembleFaceGrad using the method AssembleFaceMatrix.
In class NonlinearForm, rename the data members 'dfi', 'fbfi', etc to
use the ending 'nfi' (from nonlinear form integrator) instead of 'bfi'.

In NonlinearForm::GetEnergy, replace the warning messages with abort
calls.

In the destructor of NonlinearForm, delete the NonlinearFormIntegrator
objects stored in 'fnfi' and 'bfnfi'.

Add a version of NonlinearForm::AddBdrFaceIntegrator that allows
restricting the integrator to a subset of the boundary.

In ParNonlinearForm::GetGradient, call abort when the form has face
terms - not implemented.

In class ParNonlinearForm, add the methods GetProlongation and
GetRestriction - they use the associated ParFiniteElementSpace.