In the copy constructor of class NURBSExtension, use deep copy for the patch topology.
Remove incorrect MFEM_ASSERT from the copy constructor of class Mesh:
vertices.Size() may not equal NumOfVertices - this is now described in
the comment before the declaration of 'vertices'.

Add the explicit qualifier to some constructors in classes Mesh and
NCMesh; without these, the following implicit construction was possible:
Mesh* --> NCMesh --> Mesh (in a friend class) which was never the
intent.

In the degree-elevation constructors of class NURBSExtension, do not
issue a warning if newOrder is less than or equal to the order of a
knot vector from the parent NURBSExtention - this is now the documented
(in doxygen) behavior for this case.