In class GridFunction:
* Add an (optional) new data field, 'Vector t_vec' that may hold
  associated true-dof data, either external or owned.
* Add two methods for const and non-const access to t_vec, called
  GetTVector().
* Add a new method, SetTVector(), which as just a shortcut for
  GetTrueDofs(GetTVector()).
* Add a new method, SetFromTVector(), which is just a shortcut for
  SetFromTrueDofs(GetTVector()).
* Add two new methods, both called MakeTRef(), similar to MakeRef()
  that (re-)construct GridFunctions that wrap external true-dof data.

All of the above functionality also works for ParGridFunctions without
modifications.

Add const qualifiers to some methods in class GridFunction and to some
parameters of these methods.

Use the shortcut 'infinity()' in more places.

A few tweaks in the implementation of the methods Update(), SetSpace(),
and MakeRef() in class GridFunction.