Try to improve comments in tbilininteg.hpp
I am not entirely confident of all these comments. After brief
inspection of the code, I think the mathematical descriptions
of the Nedelec mass and curl kernels are correct, I think with
lower confidence that the descriptions of the Raviart-Thomas
mass kernels are correct, and I think the descriptions of the
RT div-div kernels are wrong (I removed those).