Add const to Data method in DenseTensor
I have an application where I pass in a const reference of a DenseTensor into a function, i.e. 
``void myFun(const DenseTensor &A) 
{
     const double * A_ptr = A.Data();
     GET_CONST_PTR(A_ptr); 
} `` 

Currently, I need access to the pointer so I can then get the device pointer from the okina memory manager. With this simple change I will be able to do that.