Reworked the 4D conforming Mesh.
Now the element gets reordered to keep the Jacobi determinant positive
and it stays positive during subsequent uniform refinement.

IMPORTANT: So far, the non-conforming 4D mesh only works when the
fix_orientation flag in the constructor of Mesh is set to false. And the
elements are thus not reordered and the Jacobi determinant can and will
be negative.