Fix bug in superlu.cpp
hypre ParCSRMatrix objects have attributes j and big_j.  In commit
84d50104 (February 2019) of the Hypre master branch, the function
hypre_MergeDiagAndOffd in hypre was changed so that the attribute
big_j is used but not j, leaving j uninitialized.  However,
superlu.cpp in mfem was not updated to account for this, so it ends
up passing a null pointer to SuperLU.  Changing superlu.cpp::144 to
use big_j instead of j seems to fix the problem.