PETSc and MFEM have a different approach in handling the Jacobians
PETSc reuses the objects, while MFEM usually gives back a new operator each time
When reusing the SNES and TS objects for different simulations,
the matrix type of the underlying DM is never update, nor is the representative matrix
for the DMSHELL (default). This is fine when using AIJ (default mattype in PETSc)
but not when using different types.