Fix compilation warnings about redefining preprocessor macros like 'MFEM_DEVICE', 'MFEM_HOST_DEVICE', etc.
In the CMake build system, set the source language to CUDA (when used)
inside the MFEM functions 'add_mfem_examples' and 'add_mfem_miniapp'.
This eliminates the need to explicitly set the language for each
example/miniapp.

In the CMake build system, ensure correct linking of all examples,
miniapps, and tests when MPI and CUDA are enabled.

When building MFEM with Umpire and CUDA/HIP, make sure Umpire is also
built with CUDA/HIP.

Small update in CHANGELOG.