Change Mesh::MakeSimplicial to use `const Mesh &`
This makes constructing a simplex mesh more convenient, allowing e.g. the one-liner

Mesh mesh = Mesh::MakeSimplicial(Mesh(mesh_file, 1, 1));

It comes at the cost of adding an ugly `const_cast`.

Note that I did not change ParMesh::MakeSimplicial. This is because a lot of the
group-related member functions of ParMesh are not const correct. Perhaps this
can be addressed in a future PR.