Handle nonconforming meshes in LOR solvers
The LOR space copies the prolongation and restriction operators from the
corresponding high-order space. This is shown to result in a spectrally
equivalent preconditioner. The DOF permutation has to be accounted for in the
prolongation operator.

Note that AMS and ADS solvers do not currently work with this approach, since
those solvers create their own finite element spaces from the mesh, and the LOR
mesh does not contain the NCMesh information needed to construct the
prolongation.