Fix repo-check history GHA history on forks.
On a fork, the repo-check history action was using the state of master
on the fork repo instead of the main repo. This made repo-check fail if
the master branch was not up-to-date.

The goal is to have branches from forks reporting repo-check correctly
when a PR is created in the main repo. This is an attempt to acheive
that.