Batched LOR: refactor
Instead of the specialized classes (H1, ND, RT) for batched LOR assembly being
derived from the abstract base class BatchedLORAssembly, they are instead
independent.

BatchedLORAssembly is now a concrete class, and it dispatches to the specialized
classes in BatchedLORAssembly::AssemblyKernel.

This is intended to reduce duplication and allow for the reuse of the
BatchedLORAssembly object