feature: adding demo container and automated builds for mfem
as discussed, this is a simple container build to provide mfem in spack, and then
show example of running examples and miniapps. This is paired with an automated build
so we always have an updated container base. This container setup can be extended to include
other kinds of builds, or other use cases

Signed-off-by: vsoch <vsoch@users.noreply.github.com>