Add another layer of AD: compute stress from derivative of scalar energy density
Puts an example of composed derivative operators. The stress is computed with
reverse mode on the scalar energy density, which is the most efficient choice.
The existing stress tangent operator is left in place, so when AD is used for
this, the 4th order tensor stress tangent operator is computed automatically
from a scalar. Neat!