Fix bug due to missing template argument
The enzyme wrapper function for the stress did not specify the
type of derivative to use for computing the stress, hence it was
defaulting to the class default (currently the symbolic stress).

I don't understand metaprogramming well enough to understand all
the consequences, but it means that the stress wrapper function
used a pointer for the wrong type when the user chose
any energy graident type other than symbolic.