Prevent the mass coefficient from being destroyed
An integrator does not own its coefficient. If a coefficient
is created in a scoped region using stack memory, then it will
expire even though a bilinear form may rely on it.

Make the mass coefficient a unique pointer owned by the Darcy
problem class. Then the coefficient will persist until a Darcy
problem instance reaches the end of the main function.