Make sure to reset size of intermediate vector
HdivSaddlePointSolver::EliminateBC was resizing the temporary vector z. When no
basis change is required, the line

basis_l2.Mult(z, xE);

would set xE = z. If z had the wrong size, this would resize xE. But xE needed
to be an alias (view) into the global x vector, so this was a bug.