Drop 'virtual' from some methods marked with both 'virtual' and 'override'
In methods related to face restriction, rename 'e_ordering' or
'ordering' to 'f_ordering' to emphasize that this is reordering
for the face dofs. Also, clarified that doxygen comments.

In class H1_ND_RT_FaceRestriction, rename some variables for better
clarity; do not use 'f_ordering' to make decisions about element
dofs.

Remove the declaration of GetFaceDofs() whose definition was
removed earlier in this branch.